modified entities  so that I can use helper functions in other data structure objects
